Add affine for-loop axiom.
It allows propagating arbitrary accumulator values in an accumulator tuple. Therefore the first application is to specialize the types for the accumulator tuple and the iter val types.
The second application then is the actual invocation where the start, stop and step are specified along with the initial values for the accumulator values and the body and return continuations. The body continuation has the current idx, the accumulator tuple and a continue cn as arguments. The continue and exit cns only the accum tuple.