Remember which imports were already performed.
This avoids symbol clashes and inconsistencies.
Since importing is transitive, a once imported dialect shall not be imported again.
Also, it is necessary to pass the current scope down when importing, since they might have to import another file as well, that has been imported further up in the hierarchy already.
This should not lead to additional name clashes, as the name clashes would otherwise appear when merging the scopes while going back up.