Start extracting mem dialect.
This already moves the respective passes to the dialect.
Since the ll backend heavily depends on mem, it is moved to the mem backend as well.
Therefore the backend is dynamically loaded as well now.
Still some work to do here, e.g. removing the mem axioms from core and adding the missing mem axioms.

This also fixes the pass dependency issue by having a map from a static ID field in the pass classes to the passes in the PassMan. If a pass is added twice, the already present instance is returned.
This obviously requires the dialects to depend on each others' passes' headers.