Add getting normalizers from dialect plugins.
The bootstrapper writes the signature of the normalizers to the dialects' header. The function names are taken from the frontend's axiom definition.
Additionally it generates an inline function `register_normalizers` that writes function ptrs to the normalizer into a axiom id->normalizer map.

This `register_normalizers` function has to be wrapped to be used in the `DialectInfo`.

The frontend registers all loaded dialects' normalizers using this callback and queries the axiom id -> function ptr map when parsing axioms.