Make DS2CPS more robust.
One major downside of this change is, that when having a loop, we must not set the PE filter to true, but without doing so, we introduce some unnecessary overhead.
Note, the loop support greatly depends on the cns being propagated through the branches as well, which currently is not implemented.