Initial BufferizationAnalysis.
This first collects all insert trees (alias sets) inside a Nest that could just work in-place in theory.
Then, this schedules (RPO) the sea of nodes into lists.
Later, uses those lists + inferred use-def information to decide which arrays must be copied due to interference.
Finally, this information is used to drive the decision whether a new buffer + copy has to be generated when bufferizing an insert.