Allow %regex.not_ and %regex.any in %regex.not_.
%regex.any can result from a mix of %regex.not_ and %regex.range.
Thus, we cannot really ensure no any will be found in a not, when constructing a not.

Equally, it is legal to use `[^012\D]`. This is a bit messed up, but in the end matches the same as `[3456789]`, i.e. first \D is negated -> [^[^0-9]] -> [0-9] and then, this is joined with [^012] which will result in [3-9].